matrix完善
