+++ Matrix.h~
@@ -1,0 +1,19 @@
+#include <iostream>
+using std::endl;
+using std::cout;
+using std::cin;
+
+class Matrix
+{
+ public:
+  Matrix(int mm, int nn)//构造函数
+  Matrix(const Matrix &src) //拷贝构造函数
+  ~Matrix()//析构函数
+  Matrix& operator=(const Matrix &src);//重载"="运算符
+  Matrix operator * (const Matrix &m2); //矩阵乘法
+  void display();
+  void input();
+ private:
+  double **data;
+  int m,n;//矩阵的行数，列数
+};
